Atividade_Auth/multi_auth/requirements.txt
Flask>=2.0
PyJWT>=2.0
Werkzeug>=2.0